feat: 통합 검색 API 연동 및 결과 렌더링 (#34)
