+++ src/app/search/page.tsx
@@ -1,66 +1,11 @@
-'use client';
-
-import { useSearchParams, useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
-import { searchByQuery } from "@/service/searchService";
-import SearchChannelResultItem from "@components/search/SearchChannelResultItem";
-import { ChannelSearchResult } from "@/types/channel";
+import { Suspense } from 'react';
 import LoadingSection from "@components/common/LoadingSection";
+import SearchContent from "@components/search";
 
 export default function SearchPage() {
-    const searchParams = useSearchParams();
-    const query = searchParams.get("q");
-    const router = useRouter();
-
-    const [channels, setChannels] = useState<ChannelSearchResult[] | null>(null);
-    const [searchType, setSearchType] = useState<string>("");
-
-    const [loading, setLoading] = useState(true);
-    useEffect(() => {
-        if (!query) return;
-
-        const fetchData = async () => {
-            try {
-                setLoading(true);
-                const result = await searchByQuery(query);
-                console.log(result);
-                setSearchType(result.searchType);
-
-                if (result.searchType === "URL") {
-                    const videoId = result.results[0]?.video.id;
-                    console.log("이동할 videoId:", videoId);
-                    if (videoId) {
-                        router.replace(`/detail/${videoId}`);
-                    }
-                } else if (result.searchType === "CHANNEL") {
-                    setChannels(result.results);
-                }
-
-            } catch (e) {
-                console.error("검색 실패:", e);
-            } finally {
-                setLoading(false);
-            }
-        };
-
-        fetchData();
-    }, [query]);
-
-    if (searchType === "URL") return null;
-    if(loading) return <LoadingSection message="검색 중입니다..." />;
-
     return (
-        <div className="max-w-screen-lg mx-auto px-4 sm:px-6 lg:px-8">
-            <h2 className="text-2xl font-bold mb-6">‘{query}’ 검색 결과</h2>
-            <div className="flex flex-col w-full gap-5">
-                {channels?.length ? (
-                    channels.map((channel) => (
-                        <SearchChannelResultItem key={channel.id} channel={channel} />
-                    ))
-                ) : (
-                    <p className="text-gray-500">검색 결과가 없습니다.</p>
-                )}
-            </div>
-        </div>
+        <Suspense fallback={<LoadingSection message="검색 페이지 로딩 중..." />}>
+            <SearchContent />
+        </Suspense>
     );
 }
